json não apaga firestore

+++ app/src/main/java/com/devminds/casasync/FirestoreHelper.kt
@@ -14,33 +14,37 @@ import com.devminds.casasync.views.HouseViewModel
 import com.devminds.casasync.views.DependentViewModel
 import com.devminds.casasync.views.TaskViewModel
 import android.content.Context
+import com.google.firebase.firestore.SetOptions
 
 object FirestoreHelper {
 
     // TODO modularizar
-    fun syncUserToFirestore(user: User) {
-        val db = FirebaseFirestore.getInstance()
-        if (user.id.isBlank()) {
-            Log.e("Firestore", "ID do usuário está nulo ou vazio. Abortando sincronização.")
-            return
-        }
+fun syncUserToFirestore(user: User) {
+    val db = FirebaseFirestore.getInstance()
 
-        val userDoc = db.collection("users").document(user.id)
+    if (user.id.isBlank()) {
+        Log.e("Firestore", "ID do usuário está nulo ou vazio. Abortando sincronização.")
+        return
+    }
 
-        // Salva dados básicos do usuário
-        val userMap = mapOf(
-            "id" to user.id,
-            "name" to user.name,
-            "login" to user.login,
-            "password" to user.password
-        )
+    val userDoc = db.collection("users").document(user.id)
 
-        userDoc.set(userMap)
-            .addOnSuccessListener { Log.d("Firestore", "Usuário sincronizado: ${user.id}") }
-            .addOnFailureListener { Log.e("Firestore", "Erro ao salvar usuário", it) }
+    // Salva dados básicos do usuário com merge para não apagar campos existentes
+    val userMap = mapOf(
+        "id" to user.id,
+        "name" to user.name,
+        "login" to user.login,
+        "password" to user.password
+    )
 
-        // Sincroniza casas
+    userDoc.set(userMap, SetOptions.merge())
+        .addOnSuccessListener { Log.d("Firestore", "Usuário sincronizado: ${user.id}") }
+        .addOnFailureListener { Log.e("Firestore", "Erro ao salvar usuário", it) }
+
+    // Sincroniza casas apenas se houver casas locais
+    if (user.houses.isNotEmpty()) {
         val casasRef = userDoc.collection("casas")
+
         casasRef.get().addOnSuccessListener { snapshot ->
             val firestoreHouseIds = snapshot.documents.map { it.id }
             val localHouseIds = user.houses.map { it.id }
@@ -59,44 +63,48 @@ object FirestoreHelper {
                 )
                 houseDoc.set(houseMap)
 
-                // Sincroniza dependentes
-                val dependentsRef = houseDoc.collection("dependentes")
-                dependentsRef.get().addOnSuccessListener { depSnapshot ->
-                    val firestoreDepIds = depSnapshot.documents.map { it.id }
-                    val localDepIds = house.dependents.map { it.id }
+                // Sincroniza dependentes apenas se houver
+                if (house.dependents.isNotEmpty()) {
+                    val dependentsRef = houseDoc.collection("dependentes")
 
-                    // Remove dependentes que não existem mais localmente
-                    firestoreDepIds.filterNot { it in localDepIds }.forEach { idToDelete ->
-                        dependentsRef.document(idToDelete).delete()
-                    }
+                    dependentsRef.get().addOnSuccessListener { depSnapshot ->
+                        val firestoreDepIds = depSnapshot.documents.map { it.id }
+                        val localDepIds = house.dependents.map { it.id }
 
-                    house.dependents.forEach { dep ->
-                        val depDoc = dependentsRef.document(dep.id)
-                        val depMap = mapOf("name" to dep.name)
-                        depDoc.set(depMap)
+                        firestoreDepIds.filterNot { it in localDepIds }.forEach { idToDelete ->
+                            dependentsRef.document(idToDelete).delete()
+                        }
 
-                        // Sincroniza tarefas
-                        val tasksRef = depDoc.collection("tarefas")
-                        tasksRef.get().addOnSuccessListener { taskSnapshot ->
-                            val firestoreTaskIds = taskSnapshot.documents.map { it.id }
-                            val localTaskIds = dep.tasks.map { it.id }
+                        house.dependents.forEach { dep ->
+                            val depDoc = dependentsRef.document(dep.id)
+                            val depMap = mapOf("name" to dep.name)
+                            depDoc.set(depMap)
 
-                            // Remove tarefas que não existem mais localmente
-                            firestoreTaskIds.filterNot { it in localTaskIds }.forEach { idToDelete ->
-                                tasksRef.document(idToDelete).delete()
-                            }
+                            // Sincroniza tarefas apenas se houver
+                            if (dep.tasks.isNotEmpty()) {
+                                val tasksRef = depDoc.collection("tarefas")
 
-                            dep.tasks.forEach { task ->
-                                val taskDoc = tasksRef.document(task.id)
-                                val taskMap = mapOf(
-                                    "name" to task.name,
-                                    "description" to task.description,
-                                    "previsionDate" to task.previsionDate,
-                                    "previsionHour" to task.previsionHour,
-                                    "startDate" to task.startDate,
-                                    "finishDate" to task.finishDate
-                                )
-                                taskDoc.set(taskMap)
+                                tasksRef.get().addOnSuccessListener { taskSnapshot ->
+                                    val firestoreTaskIds = taskSnapshot.documents.map { it.id }
+                                    val localTaskIds = dep.tasks.map { it.id }
+
+                                    firestoreTaskIds.filterNot { it in localTaskIds }.forEach { idToDelete ->
+                                        tasksRef.document(idToDelete).delete()
+                                    }
+
+                                    dep.tasks.forEach { task ->
+                                        val taskDoc = tasksRef.document(task.id)
+                                        val taskMap = mapOf(
+                                            "name" to task.name,
+                                            "description" to task.description,
+                                            "previsionDate" to task.previsionDate,
+                                            "previsionHour" to task.previsionHour,
+                                            "startDate" to task.startDate,
+                                            "finishDate" to task.finishDate
+                                        )
+                                        taskDoc.set(taskMap)
+                                    }
+                                }
                             }
                         }
                     }
@@ -104,4 +112,6 @@ object FirestoreHelper {
             }
         }
     }
+}
+
 }
